commit-iteration003

+++ ftracker.go
@@ -105,8 +105,8 @@ func WalkingSpentCalories(action int, duration, weight, height float64) float64 
 	// ваш код здесь
 
 	// ((0.035 * ВесСпортсменаВКг + (СредняяСкоростьВМетрахВСекунду**2 / РостВМетрах) * 0.029 * ВесСпортсменаВКг) * ВремяТренировкиВЧасах * minInH)
-	//return ( ( ( (walkingCaloriesWeightMultiplier * weight + (float64(action) * float64(lenStep) / (duration * 3600.0))**2 / height)) * walkingSpeedHeightMultiplier * weight) * duration * float64(minInH) )
-	return (((walkingCaloriesWeightMultiplier*weight + math.Pow((float64(action)*float64(lenStep)/(duration*3600.0)), 2)/height) * walkingSpeedHeightMultiplier * weight) * duration * float64(minInH))
+
+	return (((float64(walkingCaloriesWeightMultiplier)*weight + math.Pow((float64(action)*float64(lenStep)/(duration*3600.0)), 2)/height) * float64(walkingSpeedHeightMultiplier) * weight) * duration * float64(minInH))
 }
 
 // Константы для расчета калорий, расходуемых при плавании.
